requirements.txt
# Data Processing & Analysis
pandas
numpy
scikit-learn
torch
transformers

# Data Visualization
matplotlib
seaborn
plotly

# AI/LLM
ollama

# Web Framework
fastapi
uvicorn
pydantic
python-dotenv
python-multipart

# Database
sqlalchemy>=2.0.0
psycopg2-binary
alembic>=1.12.0
greenlet>=2.0.1
asyncpg  # For async PostgreSQL support